requirements.txt
streamlit
pandas
google-cloud-bigquery
sentence-transformers
transformers
faiss-cpu